requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
fastapi==0.110.0
filelock==3.13.1
flake8==7.0.0
fsspec==2024.3.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytube==15.0.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
